Implement different fix for circular import
